refined the map kernel structure
