+++ src/runtime/local/kernels/MAP_EXTERNAL/CtypesMapKernel_sharedMem_Pointer.py
@@ -20,7 +20,7 @@
 # Modifications Copyright 2022 The DAPHNE Consortium
 #
 # -------------------------------------------------------------
-
+import MapKernelUtils
 import numpy as np
 import ctypes
 from sympy import symbols, lambdify, sympify, Symbol
@@ -28,11 +28,11 @@ import re
 
 def apply_map_function(res_ptr, arg_ptr, rows, cols, func, varName, dtype_arg, dtype_res):
     res_array = np.ctypeslib.as_array(
-        ctypes.cast(res_ptr, ctypes.POINTER(get_ctypes_type(dtype_res))),
+        ctypes.cast(res_ptr, ctypes.POINTER(MapKernelUtils.get_ctypes_type(dtype_res))),
         shape=(rows, cols)
     )
     arg_array = np.ctypeslib.as_array(
-        ctypes.cast(arg_ptr, ctypes.POINTER(get_ctypes_type(dtype_arg))),
+        ctypes.cast(arg_ptr, ctypes.POINTER(MapKernelUtils.get_ctypes_type(dtype_arg))),
         shape=(rows, cols)
     )
 
@@ -59,22 +59,3 @@ def apply_map_function(res_ptr, arg_ptr, rows, cols, func, varName, dtype_arg, d
             res_array[:] = func_lambda(arg_array)
         except Exception as e:
             print(f"Failed to execute lambda expression: {str(e)}")
-
-def get_ctypes_type(dtype_str):
-    """Get the corresponding ctypes type for a dtype represented by a string."""
-    if dtype_str == "float32":
-        return ctypes.c_float
-    elif dtype_str == "float64":
-        return ctypes.c_double
-    elif dtype_str == "int32":
-        return ctypes.c_int32
-    elif dtype_str == "int64":
-        return ctypes.c_int64
-    elif dtype_str == "int8":
-        return ctypes.c_int8
-    elif dtype_str == "uint64":
-        return ctypes.c_uint64
-    elif dtype_str == "uint8":
-        return ctypes.c_uint8
-    else:
-        raise ValueError(f"Unsupported dtype: {dtype_str}")
