+++ src/runtime/local/kernels/MAP_EXTERNAL/CtypesMapKernel_sharedMem_voidPointer.py
@@ -1,3 +1,26 @@
+# -------------------------------------------------------------
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+# Modifications Copyright 2022 The DAPHNE Consortium
+#
+# -------------------------------------------------------------
+import MapKernelUtils
 import numpy as np
 import ctypes
 from sympy import symbols, sympify, lambdify, Symbol
@@ -5,12 +28,12 @@ import re
 
 def apply_map_function(arg_matrix_ptr, res_matrix_ptr, arg_shape, res_shape, func, varName, dtype):
     
-    ctypes_type = get_ctypes_type(dtype)
+    ctypes_type = MapKernelUtils.get_ctypes_type(dtype)
     arg_buffer = (ctypes_type * (arg_shape[0] * arg_shape[1])).from_address(arg_matrix_ptr)
     res_buffer = (ctypes_type * (res_shape[0] * res_shape[1])).from_address(res_matrix_ptr)
     
-    arg_matrix = np.frombuffer(arg_buffer, dtype=get_type(dtype)).reshape(arg_shape)
-    res_matrix = np.frombuffer(res_buffer, dtype=get_type(dtype)).reshape(res_shape)
+    arg_matrix = np.frombuffer(arg_buffer, dtype=MapKernelUtils.get_numpy_type(dtype)).reshape(arg_shape)
+    res_matrix = np.frombuffer(res_buffer, dtype=MapKernelUtils.get_numpy_type(dtype)).reshape(res_shape)
 
     match = re.search(r'def (\w+)', func)
     if match:
@@ -35,41 +58,3 @@ def apply_map_function(arg_matrix_ptr, res_matrix_ptr, arg_shape, res_shape, fun
             res_matrix[:] = func_lambda(arg_matrix)
         except Exception as e:
             print(f"Failed to execute lambda expression: {str(e)}")
-
-def get_type(dtype_str):
-    """Get the corresponding numpy type for a dtype represented by a string."""
-    if dtype_str == "float32":
-        return np.float32
-    elif dtype_str == "float64":
-        return np.double
-    elif dtype_str == "int32":
-        return np.int32
-    elif dtype_str == "int64":
-        return np.int64
-    elif dtype_str == "int8":
-        return np.int8
-    elif dtype_str == "uint64":
-        return np.uint64
-    elif dtype_str == "uint8":
-        return np.uint8
-    else:
-        raise ValueError(f"Unsupported dtype: {dtype_str}")
-    
-def get_ctypes_type(dtype_str):
-    """Get the corresponding ctypes type for a dtype represented by a string."""
-    if dtype_str == "float32":
-        return ctypes.c_float
-    elif dtype_str == "float64":
-        return ctypes.c_double
-    elif dtype_str == "int32":
-        return ctypes.c_int32
-    elif dtype_str == "int64":
-        return ctypes.c_int64
-    elif dtype_str == "int8":
-        return ctypes.c_int8
-    elif dtype_str == "uint64":
-        return ctypes.c_uint64
-    elif dtype_str == "uint8":
-        return ctypes.c_uint8
-    else:
-        raise ValueError(f"Unsupported dtype: {dtype_str}")
